+++ src/tests/inductive-proofs.ts
@@ -11,6 +11,7 @@ import { tic, toc } from '../examples/zkapps/tictoc.js';
 await isReady;
 
 let MaxProofsVerifiedZero = ZkProgram({
+  name: 'no-recursion',
   publicInput: Field,
 
   methods: {
@@ -25,6 +26,7 @@ let MaxProofsVerifiedZero = ZkProgram({
 });
 
 let MaxProofsVerifiedOne = ZkProgram({
+  name: 'recursive-1',
   publicInput: Field,
 
   methods: {
@@ -48,6 +50,7 @@ let MaxProofsVerifiedOne = ZkProgram({
 });
 
 let MaxProofsVerifiedTwo = ZkProgram({
+  name: 'recursive-2',
   publicInput: Field,
 
   methods: {